+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Current development]
 * Added new functions date_tz and strtotime_tz
+* Rework GetAmazonCustomerProfileElement to read permissions from Amazon Platform Config
 
 
 ## [Releases]