+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Releases]
 
+## 0.22.10.1 - 2022-02-01
+* Added pre-dispatch flow to Session Start block
+* Rename `Convo\Core\Eval` namespace to `Convo\Core\Expression`
 ## 0.22.10.0 - 2022-01-20
 * Added a new block role - error handler
 * Added a new flow for conversation blocks - pre-dispatch